cam-reader/requirements.txt
numpy
opencv-python
imagezmq
dotenv
paho-mqtt